pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mergify-pre-commit"
version = "1.0.0"
description = "Pre-commit hook for validating Mergify configuration files"
authors = [
    { name = "Carlos Santos" }
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "jsonschema>=4.0.0",
    "PyYAML>=5.0.0",
    "requests>=2.25.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/crsantos/mergify-pre-commit"
Repository = "https://github.com/crsantos/mergify-pre-commit"

[project.scripts]
mergify = "mergify_hook.validate_mergify:main"

[tool.setuptools.packages.find]
exclude = ["tests*"]